+++ packages/components/atoms/f-button/src/components/_tests/Button.test.js
@@ -12,6 +12,48 @@ describe('Button', () => {
 
     describe('methods ::', () => {
         describe('validateProps ::', () => {
+            let spy;
+
+            beforeEach(() => {
+                spy = jest.spyOn(global.console, 'error').mockImplementation(() => { });
+            });
+            afterEach(() => {
+                spy.mockRestore();
+            });
+
+            it.each([
+                'xsmall',
+                'small',
+                'medium',
+                'large'
+            ])('should not throw an error when buttonSize is set to %s', buttonSize => {
+                // Arrange
+                const propsData = {
+                    buttonSize,
+                    isIcon: false
+                };
+
+                // Act & Assert
+                expect(() => {
+                    shallowMount(FButton, { propsData });
+                })
+                    .not.toThrowError();
+            });
+
+            it('should throw an error when buttonSize is set to an invalid type', () => {
+                // Arrange
+                const propsData = {
+                    buttonSize: 'invalid_value',
+                    isIcon: false
+                };
+
+                // Act & Assert
+                expect(() => {
+                    shallowMount(FButton, { propsData });
+                })
+                    .toThrowError('buttonSize is set to "invalid_value"');
+            });
+
             it.each([
                 'primary',
                 'secondary',
@@ -25,29 +67,25 @@ describe('Button', () => {
                     isIcon: false
                 };
 
-                // Act
+                // Act & Assert
                 expect(() => {
                     shallowMount(FButton, { propsData });
                 })
-                    // Assert
                     .not.toThrowError();
             });
+
             it('should throw an error when the component is a standard button (isIcon="false") and buttonType is set to an invalid type', () => {
                 // Arrange
-                const spy = jest.spyOn(global.console, 'error').mockImplementation(() => { });
                 const propsData = {
                     buttonType: 'invalid_value',
                     isIcon: false
                 };
 
-                // Act
+                // Act & Assert
                 expect(() => {
                     shallowMount(FButton, { propsData });
                 })
-                    // Assert
                     .toThrowError('button is set to have buttonType="invalid_value"');
-
-                spy.mockRestore();
             });
 
             it.each([
@@ -62,30 +100,25 @@ describe('Button', () => {
                     isIcon: true
                 };
 
-                // Act
+                // Act & Assert
                 expect(() => {
                     shallowMount(FButton, { propsData });
                 })
-                    // Assert
                     .not.toThrowError();
             });
 
             it('should throw an error when the component is an iconButton (isIcon="true") and the buttonType is set to an invalid type', () => {
                 // Arrange
-                const spy = jest.spyOn(global.console, 'error').mockImplementation(() => { });
                 const propsData = {
                     buttonType: 'invalid_value',
                     isIcon: true
                 };
 
-                // Act
+                // Act & Assert
                 expect(() => {
                     shallowMount(FButton, { propsData });
                 })
-                    // Assert
                     .toThrowError('iconButton is set to have buttonType="invalid_value"');
-
-                spy.mockRestore();
             });
         });
     });